refactor: Update .gitignore, mocha.opts, babel.config.js, and .eslintrc.js
This commit updates the .gitignore file to remove the node_modules/ directory from version control. It also adds the mocha.opts file with configuration options for running tests with Mocha. The babel.config.js file is updated to use the latest stable version of @babel/preset-env and targets the current version of Node.js. Lastly, the .eslintrc.js file is updated to disable the jest environment and enable the mocha environment for running tests. These changes improve code organization and ensure compatibility with the latest dependencies and testing frameworks.